+++ src/EthernetWebServer-impl.h
@@ -12,7 +12,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.8.4
+  Version: 1.8.5
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -26,6 +26,7 @@
   1.8.2   K Hoang      27/12/2021 Fix wrong http status header bug
   1.8.3   K Hoang      28/12/2021 Fix authenticate issue caused by libb64
   1.8.4   K Hoang      11/01/2022 Fix libb64 compile error for ESP8266
+  1.8.5   K Hoang      11/01/2022 Restore support to AVR Mega2560 and add megaAVR boards. Fix libb64 fallthrough compile warning
  *************************************************************************************************************************************/
 
 #pragma once
@@ -464,6 +465,7 @@ void EthernetWebServer::_prepareHeader(String& response, int code, const char* c
   _responseHeaders = String("");
 }
 
+#if ! ( ETHERNET_USE_AVR_MEGA || ETHERNET_USE_MEGA_AVR )
 void EthernetWebServer::_prepareHeader(EWString& response, int code, const char* content_type, size_t contentLength) 
 {
   response = "HTTP/1." + fromString(String(_currentVersion)) + " ";
@@ -507,6 +509,7 @@ void EthernetWebServer::_prepareHeader(EWString& response, int code, const char*
   
   _responseHeaders = String("");
 }
+#endif
 
 void EthernetWebServer::send(int code, const char* content_type, const String& content)
 {